SCRUM-1138 branch has typo, it should be this ticket.  Make a new Visit Mod section to put below resources, take the Visit Mod link out of the About and add it to the h2 header of the new Visit Mod section
