+++ apps/main-app/src/containers/modLanding/About.js
@@ -1,15 +1,13 @@
 import React from 'react';
 import style from './style.scss';
 import PropTypes from 'prop-types';
-import LinkToMOD from './LinkToMOD';
 
-const About = ({htmlContent, modVisitButtonText, linkToMod, sectionStyle}) => {
+const About = ({htmlContent, sectionStyle}) => {
   return (
     <div className={`container ${style.containerExtra}`}>
       <div data-testid="about-inner-container" className={`${style.section} ${sectionStyle}`}>
         <h2 data-testid="header" className={style.sectionTitle}>About</h2>
         <div data-testid="html-content" dangerouslySetInnerHTML={{__html: htmlContent}}/>
-        <LinkToMOD modName={modVisitButtonText} linkAddress={linkToMod}/>
       </div>
     </div>
   );
@@ -17,8 +15,6 @@ const About = ({htmlContent, modVisitButtonText, linkToMod, sectionStyle}) => {
 
 About.propTypes = {
   htmlContent: PropTypes.string.isRequired,
-  modVisitButtonText: PropTypes.string.isRequired,
-  linkToMod: PropTypes.string.isRequired
 }
 
 export default About;
